agent/requirements.txt
feedparser>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
openai>=1.0.0
openai-agents>=0.0.7
pydantic>=2.0.0
